refactor: remove abstract AnonCreds* classes, rename AnonCredsRs* classes
