+++ aries_cloudagent/anoncreds/anoncreds/routes.py
@@ -12,7 +12,7 @@ from aiohttp_apispec import (
 from marshmallow import fields
 
 from .registry import AnonCredsRegistry
-from ..issuer import AnonCredsIssuer
+from .issuer import AnonCredsIssuer
 from ..models.anoncreds_cred_def import (
     CredDefSchema,
     GetCredDefResultSchema,
