+++ aries_cloudagent/anoncreds/models/xform.py
@@ -1,6 +1,6 @@
 """Utilities to deal with indy."""
 
-from ...anoncreds.holder import AnonCredsHolder
+from ...anoncreds.anoncreds.holder import AnonCredsHolder
 
 from .pres_preview import IndyPresPreview
