+++ aries_cloudagent/protocols/issue_credential/v1_0/handlers/credential_offer_handler.py
@@ -3,7 +3,7 @@
 
 from .....wallet.util import default_did_from_verkey
 from .....core.oob_processor import OobMessageProcessor
-from .....anoncreds.holder import AnonCredsHolderError
+from .....anoncreds.anoncreds.holder import AnonCredsHolderError
 from .....ledger.error import LedgerError
 from .....messaging.base_handler import BaseHandler, HandlerException
 from .....messaging.models.base import BaseModelError
