+++ aries_cloudagent/protocols/issue_credential/v1_0/routes.py
@@ -16,8 +16,8 @@ from ....wallet.util import default_did_from_verkey
 from ....admin.request_context import AdminRequestContext
 from ....connections.models.conn_record import ConnRecord
 from ....core.profile import Profile
-from ....anoncreds.holder import AnonCredsHolderError
-from ....anoncreds.issuer import AnonCredsIssuerError
+from ....anoncreds.anoncreds.holder import AnonCredsHolderError
+from ....anoncreds.anoncreds.issuer import AnonCredsIssuerError
 from ....ledger.error import LedgerError
 from ....messaging.credential_definitions.util import CRED_DEF_TAGS
 from ....messaging.models.base import BaseModelError
