+++ aries_cloudagent/protocols/issue_credential/v2_0/formats/indy/handler.py
@@ -8,8 +8,8 @@ from typing import Mapping, Tuple
 from marshmallow import RAISE
 
 from ......anoncreds.anoncreds.registry import AnonCredsRegistry
-from ......anoncreds.holder import AnonCredsHolder, AnonCredsHolderError
-from ......anoncreds.issuer import (
+from ......anoncreds.anoncreds.holder import AnonCredsHolder, AnonCredsHolderError
+from ......anoncreds.anoncreds.issuer import (
     AnonCredsIssuer,
     AnonCredsIssuerRevocationRegistryFullError,
 )
