+++ aries_cloudagent/protocols/issue_credential/v2_0/handlers/cred_offer_handler.py
@@ -2,7 +2,7 @@
 
 from .....wallet.util import default_did_from_verkey
 from .....core.oob_processor import OobMessageProcessor
-from .....anoncreds.holder import AnonCredsHolderError
+from .....anoncreds.anoncreds.holder import AnonCredsHolderError
 from .....ledger.error import LedgerError
 from .....messaging.base_handler import BaseHandler, HandlerException
 from .....messaging.models.base import BaseModelError
