+++ aries_cloudagent/protocols/issue_credential/v2_0/handlers/cred_request_handler.py
@@ -1,7 +1,7 @@
 """Credential request message handler."""
 
 from .....core.oob_processor import OobMessageProcessor
-from .....anoncreds.issuer import AnonCredsIssuerError
+from .....anoncreds.anoncreds.issuer import AnonCredsIssuerError
 from .....ledger.error import LedgerError
 from .....messaging.base_handler import BaseHandler, HandlerException
 from .....messaging.models.base import BaseModelError
