+++ aries_cloudagent/protocols/present_proof/indy/pres_exch_handler.py
@@ -5,7 +5,7 @@ import time
 from typing import Union
 
 from ....anoncreds.anoncreds.registry import AnonCredsRegistry
-from ....anoncreds.holder import AnonCredsHolder, AnonCredsHolderError
+from ....anoncreds.anoncreds.holder import AnonCredsHolder, AnonCredsHolderError
 from ....anoncreds.models.xform import indy_proof_req2non_revoc_intervals
 from ....core.error import BaseError
 from ....core.profile import Profile
