+++ aries_cloudagent/protocols/present_proof/v1_0/routes.py
@@ -14,7 +14,7 @@ from marshmallow import fields, validate
 
 from ....admin.request_context import AdminRequestContext
 from ....connections.models.conn_record import ConnRecord
-from ....anoncreds.holder import AnonCredsHolder, AnonCredsHolderError
+from ....anoncreds.anoncreds.holder import AnonCredsHolder, AnonCredsHolderError
 from ....anoncreds.models.cred_precis import IndyCredPrecisSchema
 from ....anoncreds.models.proof import IndyPresSpecSchema
 from ....anoncreds.models.proof_request import IndyProofRequestSchema
