+++ aries_cloudagent/protocols/present_proof/v1_0/tests/test_routes.py
@@ -5,9 +5,9 @@ from asynctest import mock as async_mock, TestCase as AsyncTestCase
 from marshmallow import ValidationError
 
 from .....admin.request_context import AdminRequestContext
-from .....anoncreds.holder import AnonCredsHolder
+from .....anoncreds.anoncreds.holder import AnonCredsHolder
 from .....anoncreds.models.proof_request import IndyProofReqAttrSpecSchema
-from .....anoncreds.verifier import AnonCredsVerifier
+from .....anoncreds.anoncreds.verifier import AnonCredsVerifier
 from .....ledger.base import BaseLedger
 from .....storage.error import StorageNotFoundError
 
@@ -142,7 +142,9 @@ class TestProofRoutes(AsyncTestCase):
             AnonCredsHolder,
             async_mock.MagicMock(
                 get_credentials_for_presentation_request_by_referent=(
-                    async_mock.CoroutineMock(side_effect=test_module.AnonCredsHolderError())
+                    async_mock.CoroutineMock(
+                        side_effect=test_module.AnonCredsHolderError()
+                    )
                 )
             ),
         )
