+++ aries_cloudagent/protocols/present_proof/v2_0/handlers/pres_request_handler.py
@@ -1,7 +1,7 @@
 """Presentation request message handler."""
 
 from .....core.oob_processor import OobMessageProcessor
-from .....anoncreds.holder import AnonCredsHolderError
+from .....anoncreds.anoncreds.holder import AnonCredsHolderError
 from .....ledger.error import LedgerError
 from .....messaging.base_handler import BaseHandler, HandlerException
 from .....messaging.models.base import BaseModelError
